Move lightweight cluster config in separate package
... and update docs
